Unbuild antipattern
